requirements.txt
flask
flask-cors
python-pptx
requests
werkzeug
sqlite3
pillow
python-dateutil